+++ src/App.jsx
@@ -1,43 +1,12 @@
 import { useState, useEffect } from "react";
 import Game from "./components/Game/Game";
-import dummyImg from "./assets/images/dummy.jpg";
 import LevelStart from "./components/LevelStart/LevelStart";
 import GameOver from "./components/GameOver/GameOver";
-
-const dummyStore = () => {
-  const store = [];
-  for (let i = 0; i < 250; i++) {
-    const obj = { name: "dummy" + i, head: i, image: dummyImg };
-    store.push(obj);
-  }
-  return store;
-};
-
-// utils
-
-const shuffleArray = (arr) => {
-  const copyArr = [...arr];
-  for (let i = copyArr.length - 1; i > 0; i--) {
-    const rand = Math.floor(Math.random() * (i + 1));
-    [copyArr[i], copyArr[rand]] = [copyArr[rand], copyArr[i]];
-  }
-  return copyArr;
-};
-
-const randomSubArray = (arr, len) => {
-  const shuffle = shuffleArray(arr);
-  return len >= shuffle.length ? shuffle : shuffle.slice(0, len);
-};
-
-const createLevel = (db, modifier) => {
-  const len = 4 + (modifier - 1) * 2;
-  return randomSubArray(db, len).map((info) => ({
-    name: info.name,
-    id: `${info.head}${info.tail}`,
-    image: info.image,
-    clicked: false,
-  }));
-};
+import {
+  randomSubArray,
+  createLevel,
+  shuffleArray,
+} from "../utils/gameManager";
 
 function App() {
   const [cardStore, setCardStore] = useState([]);
